+++ src/td/stage.h
@@ -80,7 +80,7 @@ class StageClass {
 
   int Fetch_Stage() const { return Stage; }
   int Fetch_Rate() const { return Rate; }
-  void Set_Stage(int stage) { Stage = stage; }
+  void Set_Stage(int stage) { Stage = static_cast<unsigned short>(stage); }
   void Set_Rate(unsigned char rate) { Rate = StageTimer = rate; }
   void AI() {}
   bool Graphic_Logic() {
